moved connect stuff
